backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# LangGraph
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data processing
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0

# ML
scikit-learn>=1.5.0
joblib>=1.4.0

# HTTP clients
requests>=2.32.0
httpx>=0.27.0

# Medical code APIs
omophub>=0.1.0

# Evaluation
matplotlib>=3.9.0
seaborn>=0.13.0
